refactor(shared): move pure pricing/transformers/types to shared
- Move pure pricing, transformers, specialTypes, and uiTypes to @costwise/shared
- Update imports across packages/domain, apps/api, and apps/web
- Relocate helper tests to packages/shared/src/helpers.test.ts
- Verification gates:
  - grep -rn '@costwise/db' packages/shared (empty)
  - grep -rn --exclude-dir='.next' 'domain/utils/pricing|domain/utils/transformers|domain/types/specialTypes' apps packages (empty)
  - turbo run build (6/6 packages green)
  - turbo run test (5/5 suites, 126 tests green)
  - turbo run lint (green)
