+++ apps/web/src/app/(user)/recipes/create/page.tsx
@@ -4,7 +4,7 @@ import { IngredientService } from '@costwise/domain/services/ingredientService';
 import { RecipeIngredients } from '@costwise/shared/recipe';
 import { getServerSession } from '@/app/lib/serverSession';
 import { redirect } from 'next/navigation';
-import { Metadata } from '@costwise/domain/types/specialTypes';
+import { Metadata } from '@costwise/shared/specialTypes';
 
 const Page = async () => {
   const session = await getServerSession();
