+++ apps/web/src/app/components/shared/table/tableHead.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { TableHeadColumn } from '@costwise/domain/types/specialTypes';
+import { TableHeadColumn } from '@costwise/shared/specialTypes';
 import SortedLink from '../sortedLink';
 import { cn } from '@/app/utils/cn';
