+++ apps/web/src/app/hooks/useRecipeForm.tsx
@@ -18,8 +18,8 @@ import { v4 as uuidv4 } from "uuid";
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { calculateRecipeData, getTotalPrice } from '@costwise/domain/utils/pricing';
-import { getArrayChanges } from '@costwise/domain/utils/transformers';
+import { calculateRecipeData, getTotalPrice } from '@costwise/shared/pricing';
+import { getArrayChanges } from '@costwise/shared/transformers';
 import { sendRecipe, sendRecipeToUpdate } from '@/app/services/services';
 import { useRouter } from 'next/navigation';
 import { useFileUpload } from './useFileUpload';
