+++ packages/domain/src/repositories/addressesRepository.ts
@@ -1,6 +1,6 @@
 import { Database, supplierAddresses } from "@costwise/db/schema";
 import { IAddressesRepository, OperationResult } from "../types/repositories";
-import { DBSupplierAddress } from "../types/specialTypes";
+import { DBSupplierAddress } from "@costwise/shared/specialTypes";
 import { eq } from "drizzle-orm";
 import { DatabaseError } from "../utils/errors";
