+++ packages/domain/src/repositories/supplierFinancialDataRepository.ts
@@ -3,7 +3,7 @@ import {
   ISupplierFinancialDataRepository,
   OperationResult,
 } from "../types/repositories";
-import { DBSupplierFinancialData } from "../types/specialTypes";
+import { DBSupplierFinancialData } from "@costwise/shared/specialTypes";
 import { eq } from "drizzle-orm";
 import { DatabaseError } from "../utils/errors";
